Addressing issue #1090

+++ libraries/TGRSIint/TGRSIint.cxx
@@ -622,9 +622,8 @@ void TGRSIint::SetupPipeline()
 	// If needed, generate the individual detectors from the TFragments
 	if(generate_analysis_data) {
 		TGRSIOptions::AnalysisOptions()->Print();
-		eventBuildingLoop = TEventBuildingLoop::Get("5_event_build_loop", event_build_mode);
+		eventBuildingLoop = TEventBuildingLoop::Get("5_event_build_loop", event_build_mode, opt->AnalysisOptions()->BuildWindow());
 		eventBuildingLoop->SetSortDepth(opt->SortDepth());
-		eventBuildingLoop->SetBuildWindow(opt->AnalysisOptions()->BuildWindow());
 		if(unpackLoop != nullptr) {
 			eventBuildingLoop->InputQueue() = unpackLoop->AddGoodOutputQueue();
 		}
@@ -708,22 +707,23 @@ bool TGRSIInterruptHandler::Notify()
 {
 	/// When ctrl-c is pressed, this takes over. This can be used in the future
 	/// for safe cleanup.
-	static int timespressed = 0;
-	timespressed++;
-	switch(timespressed) {
+	if(!StoppableThread::AnyThreadRunning()) {
+		std::cout<<std::endl<<DRED<<BG_WHITE<<"   Control-c was pressed in interactive mode.   "<<RESET_COLOR<<std::endl;
+		exit(1);
+	}
+	static int timesPressed = 0;
+	timesPressed++;
+	switch(timesPressed) {
 		case 1:
-			printf("\n" DRED BG_WHITE "   Control-c was pressed, terminating input loop.   " RESET_COLOR "\n");
-			fflush(stdout);
+			std::cout<<std::endl<<DRED<<BG_WHITE<<"   Control-c was pressed, terminating input loop.   "<<RESET_COLOR<<std::endl;
 			TGRSIint::instance()->Terminate();
 			break;
 		case 2:
-			printf("\n" DRED BG_WHITE "   Control-c was pressed again, stopping all queues.   " RESET_COLOR "\n");
-			fflush(stdout);
+			std::cout<<std::endl<<DRED<<BG_WHITE<<"   Control-c was pressed, stopping all queues.   "<<RESET_COLOR<<std::endl;
 			StoppableThread::ClearAllQueues();
 			break;
 		default:
-			printf("\n" DRED BG_WHITE "   No you shutup!   " RESET_COLOR "\n");
-			fflush(stdout);
+			std::cout<<std::endl<<DRED<<BG_WHITE<<"   No you shutup!   "<<RESET_COLOR<<std::endl;
 			exit(1);
 	}
 	return true;
